refactor : pojo 폴더 생성으로 파일 구조 수정
